Turn the module into a package
